+++ src/scriptlets/scriptlets-list.js
@@ -73,6 +73,7 @@ export * from './trusted-suppress-native-method';
 export * from './json-prune-xhr-response';
 export * from './trusted-dispatch-event';
 export * from './trusted-replace-outbound-text';
+export * from './prevent-canvas';
 // redirects as scriptlets
 // https://github.com/AdguardTeam/Scriptlets/issues/300
 export * from './amazon-apstag';